feat: 데이터 없을때 view용 그래프 GraphUtils로 분리
코드가 너무 길어 데이터 없을때 view용 그래프 GraphUtils로 분리

issue: #40
